Move nftOperationId to coin-framework
